+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ##### Bug Fixes
 
 - Unused parameter warnings are suppressed in `@available(*, unavailable)` functions.
+- Fix handling of Xcode projects with single and double quotes in their path and scheme names.
 
 ## 3.0.3 (2025-03-15)